requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.3
pg8000==1.31.2
Flask-Migrate==4.0.4
python-dotenv==1.0.0
Faker==18.9.0
